Threads: Move clear-TID metadata into Rust
Rust now consumes the authoritative ThreadState on exit and returns its
CLONE_CHILD_CLEARTID pointer to the shared host runtime. The host keeps
the process-memory clear and futex wake because it owns guest memory.

Use an i64 export result so every wasm32 pointer remains nonnegative and
negative values truthfully carry errno. This removes the duplicate
TypeScript pointer map without weakening exact task or retry cleanup.

Validation:
- ABI snapshot and compatibility check
- kernel wasm release build
- 2 focused kernel thread-exit tests
- host type and bundle builds
- 86 clone, exec, entry-gate, and lifecycle tests
- 3 focused process-retirement tests
- 10 generated-ABI tests

+++ host/test/support/kernel-scratch-instance.ts
@@ -435,7 +435,7 @@ function signatures(
     },
     kernel_thread_exit: {
       parameters: [i32, i32],
-      result: i32,
+      result: i64,
     },
     kernel_thread_has_deliverable: {
       parameters: [i32, i32],
